Move groupEntityFactory to group module
