+++ apps/server/src/modules/board/controller/api-test/board-update-title-in-room.api.spec.ts
@@ -1,15 +1,15 @@
 import { EntityManager } from '@mikro-orm/mongodb';
+import { accountFactory } from '@modules/account/testing';
+import { GroupEntityTypes } from '@modules/group/entity';
+import { groupEntityFactory } from '@modules/group/testing/group-entity.factory';
+import { roomMembershipEntityFactory } from '@modules/room-membership/testing';
+import { roomEntityFactory } from '@modules/room/testing';
 import { ServerTestModule } from '@modules/server/server.module';
 import { INestApplication } from '@nestjs/common';
 import { Test, TestingModule } from '@nestjs/testing';
 import { ApiValidationError } from '@shared/common';
 import { Permission, RoleName } from '@shared/domain/interface';
-import { accountFactory } from '@src/modules/account/testing';
-import { GroupEntityTypes } from '@src/modules/group/entity';
-import { roomMembershipEntityFactory } from '@src/modules/room-membership/testing';
-import { roomEntityFactory } from '@src/modules/room/testing';
 import { cleanupCollections } from '@testing/cleanup-collections';
-import { groupEntityFactory } from '@testing/factory/group-entity.factory';
 import { roleFactory } from '@testing/factory/role.factory';
 import { userFactory } from '@testing/factory/user.factory';
 import { TestApiClient } from '@testing/test-api-client';
