+++ apps/server/src/modules/group/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './group-entity.factory';
