+++ apps/server/src/modules/room/api/test/room-delete.api.spec.ts
@@ -1,13 +1,13 @@
 import { EntityManager, ObjectId } from '@mikro-orm/mongodb';
 import { GroupEntityTypes } from '@modules/group/entity/group.entity';
+import { groupEntityFactory } from '@modules/group/testing';
+import { RoomMembershipEntity } from '@modules/room-membership';
+import { roomMembershipEntityFactory } from '@modules/room-membership/testing/room-membership-entity.factory';
 import { ServerTestModule, serverConfig, type ServerConfig } from '@modules/server';
 import { HttpStatus, INestApplication, NotFoundException } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { Permission, RoleName } from '@shared/domain/interface';
-import { RoomMembershipEntity } from '@src/modules/room-membership';
-import { roomMembershipEntityFactory } from '@src/modules/room-membership/testing/room-membership-entity.factory';
 import { cleanupCollections } from '@testing/cleanup-collections';
-import { groupEntityFactory } from '@testing/factory/group-entity.factory';
 import { roleFactory } from '@testing/factory/role.factory';
 import { schoolEntityFactory } from '@testing/factory/school-entity.factory';
 import { UserAndAccountTestFactory } from '@testing/factory/user-and-account.test.factory';
