+++ apps/server/src/modules/tool/tool-launch/service/auto-parameter-strategy/auto-group-external-uuid.strategy.spec.ts
@@ -4,12 +4,12 @@ import { BoardExternalReferenceType, BoardNodeService } from '@modules/board';
 import { columnBoardFactory, externalToolElementFactory } from '@modules/board/testing';
 import { Group, GroupService } from '@modules/group';
 import { GroupEntity } from '@modules/group/entity';
+import { groupEntityFactory } from '@modules/group/testing';
 import { CourseService } from '@modules/learnroom';
 import { Test, TestingModule } from '@nestjs/testing';
 import { Course } from '@shared/domain/entity';
 import { courseFactory } from '@testing/factory/course.factory';
 import { groupFactory } from '@testing/factory/domainobject';
-import { groupEntityFactory } from '@testing/factory/group-entity.factory';
 import { setupEntities } from '@testing/setup-entities';
 import { ToolContextType } from '../../../common/enum';
 import { ContextExternalTool } from '../../../context-external-tool/domain';
